Use fully qualified resource id for notification icon
With the update to the cordova-platform 6.1 the app icon has changed
from a drawable to a mipmap. Therefore we can not substitute the
ANDROID_NOTIFICATION_ICON drawable resource name anymore since
the default value needs to point to a mipmap resource whereas the regular
app specific notification icon will be stored as a drawable resource.

Therefore the developer now has to provide a fully qualified resource name
as the ANDROID_NOTIFICATION_ICON.

Example:

<plugin name="tabris-plugin-firebase" spec="1.0.0">
  <variable name="ANDROID_NOTIFICATION_ICON" value="@drawable/ic_notification" />
</plugin>