Simplify Firebase credential configuration
Expect Firebase credential file on iOS to be in cordova/ and not in
cordova/res/ios. This is the same location as the one expected for the
Android credential file.

Remove example GoogleService-Info.plist file. The user should provide
his own credential files in cordova/, otherwise no reasonable results
are expected. There is no example credentials file for Android either.

Fix README stating that the credential files should be placed in the
example root. This was most likely true for the pre-Tabris.js CLI
Cordova-based app folder structure.

Change-Id: I70551d76f0098b097ec48d477bc1c74453f0c2ae