Improve misleading error message
The error message shown when the google-services.json file is not found 
is misleading, since the file should be located in an app subdirectory named
/cordova and not in its root.

Since Tabris.js 2.x stable, Cordova-based project layout is not supported 
anymore. Cordova artifacts will be placed in a /cordova subdirectory in the
root of the app and "context.opts.projectRoot" refered to in scripts will
resolve with its path.