Update supported Tabris version to 2.x
Since custom widget API is final now, it should not break between minor
version updates of Tabris.js.

Use carret version syntax for the Cordova plugin as this is the default
semver specifier used by npm. This will provide non-breaking
improvements to the plugin and prevent having to update the version in
the README on each patch version update.

Change-Id: Iaee7e7bcba84efb9462c264bf2f07ec0c8c8e516