Fix broken AppDelegate after injecting the sources
It's possible that methods implemented by this plugin are already
implemented in AppDelegate.m file. Using the script with its current
state leads to a compile-time error:
"Duplicate implementation for a method: xxx"

We have to check first if the method is implemented, and if it is then
we inject the new code to the existing implementation. If it's not then
we just put in the whole method. Like we did before.