Support integration with the Google Tag Manager
Introduce a FEATURES variable as part of a modular build, which will
strip feature declarations from plugin.xml to prevent their code being
compiled and run. To be extended to analytics and cloud messaging in a
follow-up (breaking) change.

Let containers reside in directories called
tagmanager-container/{{platform}} in order for the plugin to be agnostic
about the naming scheme of the container.

Change-Id: I1d9ddb9fc4afa0c96eede83bf5082fe724dab001