Fix 'plugin with id not found' error
When we build a Tabris.js app with the 'tabris-plugin-firebase' plugin, the build fails with the error "Plugin with id 'com.google.gms.google-services' not found.".

The change uses the class name GoogleServicesPlugin instead of id(string) to be able to apply the plugin from the non-root gradle file, which solves the issue. Additionally, the change applies the plugin if the `googleServices` is not added to `cdvPluginPostBuildExtras` somewhere else.

Fixes #71