Upgrade Android com.google.firebase dependencies
On Android 12 (API Level 31) and higher incoming push notification are
not shown in the notification bar any more. Fix this by upgrading all
Firebase Android dependencies to the latest one compatible with the
internal implementation according to the BoM (Bill of Meterials) 30.5.0
from 2022-09-15.

https://firebase.google.com/support/release-notes/android#2022-09-15

Also add an explicit dependency to firebase-iid to provide a compatible
implementation for the instance ID that is still in use in some places.