Update example for Tabris 3.9 nightly
Add `cordova/package.json` to the example which will be required as of
Tabris.js 3.9.0.

Also move example from `.example` back to `example`. Having the example
in `.example` was a workaround for apache/cordova-cli#401 that does not
work anymore. Workaround this issue by using a Cordova hook instead.

Configure `android:usesCleartextTraffic="true"` to enable side-loading
the example (see eclipsesource/tabris-js#1857).