Adapt plugin to notification trampoline restrictions
Applications that target Android 12 or higher can’t start activities
from services or broadcast receivers that are used as a notification
trampolines. After the user taps on a notification or an action button
within the notification, the app cannot call `startActivity()` inside
of a service or broadcast receiver which helps to improve the app
performance and user experience.

Replace the NotificationOpenedReceiver with a PendingIntent object that
is associated with the activity that users see after they tap on the
notification.

https://developer.android.com/about/versions/12/behavior-changes-12#notification-trampolines